support for deep links as login qr codes
